Time: 44 ms (78.77%), Space: 27.1 MB (35.89%) - LeetHub

+++ 1663-smallest-string-with-a-given-numeric-value/1663-smallest-string-with-a-given-numeric-value.cpp
@@ -3,7 +3,7 @@ public:
     string getSmallestString(int n, int k) {
         
         string ans;
-        long int i;
+        int i;
         
         k=k-n;
